requirements.txt
pymysql==1.1.0
cryptography==41.0.5
python-dotenv==1.0.0
tabulate==0.9.0
colorama==0.4.6
